feat: Add GET /okr endpoint
