+++ package.json
@@ -33,12 +33,14 @@
   "devDependencies": {
     "@types/express": "4.17.13",
     "@types/jest": "27.4.0",
-    "@typescript-eslint/eslint-plugin": "^5.10.1",
-    "@typescript-eslint/parser": "^5.10.1",
+    "@types/supertest": "2.0.11",
+    "@typescript-eslint/eslint-plugin": "5.10.1",
+    "@typescript-eslint/parser": "5.10.1",
     "babel-jest": "27.4.6",
     "eslint": "8.8.0",
     "jest": "27.4.7",
     "nodemon": "2.0.15",
+    "supertest": "6.2.2",
     "ts-jest": "27.1.3",
     "ts-node": "10.4.0",
     "typescript": "4.5.5"
