+++ src/components/okr/aplication/getOkrController.ts
@@ -1,0 +1,9 @@
+import { Request, Response } from 'express';
+
+export const getOkrController = async (req: Request, res: Response) => {
+  try {
+    res.status(200).send('ok');
+  } catch (error) {
+    res.status(500).send('error');
+  }
+};
