+++ src/components/okr/aplication/index.ts
@@ -1,0 +1,1 @@
+export * from './getOkrController';
