--- src/routes/routes.ts
@@ -1,3 +1,0 @@
-export const suma = (a: number, b: number) => {
-    return a + b;
-};
